requirements.txt
Flask
pytesseract
pillow
pdf2image
gunicorn
